invoice
Update motorhome fix
pricehandler
homecontroller getmapping for invoice
fullprice field i contract

+++ src/main/java/com/nmr/Handler/PriceHandler.java
@@ -5,9 +5,10 @@ import java.time.LocalDateTime;
 import static java.time.temporal.ChronoUnit.DAYS;
 
 public class PriceHandler {
-    public static int calculateFullPrice(LocalDateTime rentalStartDate, LocalDateTime rentalEndDate, int pricePerDay){
+
+    public static double calculateFullPrice(LocalDateTime rentalStartDate, LocalDateTime rentalEndDate, int pricePerDay){
         long days = DAYS.between(rentalStartDate, rentalEndDate);
         double fullPrice = days * pricePerDay;
-        return (int)fullPrice;
+        return fullPrice;
     }
 }
